src/fabric_de_mcp/requirements.txt
azure-identity
requests
mcp
python-dotenv
